+++ cmd/feedfinder.go
@@ -1,10 +1,11 @@
-// Copyright (c) 2020, SpecializedGeneralist Authors. All rights reserved.
+// Copyright (c) 2022, SpecializedGeneralist Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
 package main
 
 import (
+	"flag"
 	"fmt"
 	"github.com/SpecializedGeneralist/feedfinder"
 	"log"
@@ -12,23 +13,20 @@ import (
 )
 
 func main() {
-	log.SetFlags(0)
+	initialURL := flag.String("url", "", "initial URL to probe")
+	maxDepth := flag.Int("depth", 2, "maximum depth of linked pages to follow (>= 1)")
+	flag.Parse()
 
-	if len(os.Args) != 2 {
-		printUsageAndExit()
+	if len(flag.Args()) != 0 {
+		flag.Usage()
+		os.Exit(1)
 	}
 
-	url := os.Args[1]
-	feeds, err := feedfinder.FindFeeds(url)
+	feeds, err := feedfinder.FindFeeds(*initialURL, *maxDepth)
 	if err != nil {
 		log.Fatal(err)
 	}
 	for _, feed := range feeds {
 		fmt.Println(feed)
 	}
-}
-
-func printUsageAndExit() {
-	log.Fatalf("feedfinder v%s\nUsage: %s <URL>",
-		feedfinder.Version, os.Args[0])
 }